[509210] Fix duplicate GUID on both nodes
The issue is caused by the value in dictionary being changed as it's
passing to register's method as reference. The guid of first node has
been overwritten by the guid of second node. To fix the issue, we pass
the copy of dictionary into register's method so that the changes on
original dictionary won't affect the registeration process.

Signed-off-by: Wong Yeng Liong <yeng.liong.wong@intel.com>