(531066) Fix TB device key/cert regression (#357)
In 4.1.1, we introduced some bug fixes discovered by fuzz testing. One of the fixes assumed that a missing device key in x.509 mode in cloudadapter was an error, so it skipped loading the cert chain in that case. In fact, when we provision with ThingsBoard, we assume both cert/key are in the same .pem file so we would only have a cert file, and the load_cert_chain library function we call allows this scenario.

In working on this bug, I discovered that thingsboard's official documentation has switched to using separate files for x.509 certs/keys, so we will need to add a story to update our documentation to be explicit as to how to generate the combined file, and warn to only paste the cert part into thingsboard when setting up a device.