Add INBS cloud + ping command to Cloudadapter (#496)
This PR adds support for the INBS cloud to cloudadapter, along with ping support, and the ability
to turn on secure connection with TLS + token. See the INBS guide for more details.

Additionally, a mock INBS server is added for testing without the cloud. The INBS server will send
pings once per second to cloudadapter.