Add firmware blob update capability using fwupdtool (#519)
* Add firmware blob update capability using fwupdtool

- Introduce UpdateFirmwareBlobFwupdtool.sh script for firmware updates
- Update firmware_tool_info.conf to use the new script
- Modify AppArmor configurations to allow script execution

This change enables firmware blob updates for Arrow Lake Client Platform
using fwupdtool.

* fix base setup build

* Fix GUID detection for case where firmware tool is fwupdmgr

When we call fwupdmgr via UpdateFirmwareBlobFwupdtool.sh, we need to
be able to detect the firmware GUID via '-l' passed in to the shell
script. Here I've given the script permission to call 'awk' and
put in logic to detect and return the guid if '-l' is passed in.

* add Changelog