[NEXMANAGE-493] Fixed dispatcher failed to get granular log during download-only mode
During download-only mode, there are no package installations. The
dispatcher agent failed to retrieve the package information from history.log.
This PR fixes the issue by instructing the dispatcher agent not to write
the granular log if it is in download-only mode. Additionally, this PR
includes the exception handling to address the corner cases in
history.log, such as when no upgrade package has been found.

Signed-off-by: yengliong <yeng.liong.wong@intel.com>