Create a `docker-compose-dev` tool for quick INBM container testing without rebuilds (#566)
This PR provides a `docker-compose` setup to run the INBM agents in containers 
with a rapid build for dev testing.  The build is quicker because it doesn't use PyInstaller—instead, it simply
copies the source files to each container.  Additionally, the agents run in insecure mode, so
no certs have to be provisioned. No apparmor rules apply, and there are no mqtt ACLs.

Cloud config is easy--simply provide `adapter.cfg` alongside
the `docker-compose-dev` files. To access a cloud service (such as INBS) running on `localhost`, simply
use the host named `host.docker.internal` in `adapter.cfg`.

This container setup does not replace the need for full native testing and should not be
run in production.