[NEXMANAGE-846] Fix granular log raise error when granular log file is empty (#572)
* [NEXMANAGE-846] Fix granular log raise error when granular log file is
empty

The recent changes in PUA no longer remove the granular log file.
Instead, it truncates the file. This causes an error in INBM because the
file exists but contains empty data. The PR resolves the issue by
checking the content of the file. If the file exists with empty data, it
will create the template and continue the package checks.

Signed-off-by: yengliong <yeng.liong.wong@intel.com>

* Handle OSError

---------

Signed-off-by: yengliong <yeng.liong.wong@intel.com>