[NEXMANAGE-737] Sota cancel mode with threading event method (#567)
* [NEXMANAGE-737] Enable sota cancel mode.

This PR implements the sota cancel mode using a threading event flag. When the thread is created, it will be added to a thread list.
When the dispatcher receives the sota cancel request, it checks the current running thread and retrieves its sota type. If it is a download-only sota, the dispatcher sets the event flag to issue a cancel request.
Please note that the first inbc command should be terminated before sending inbc cancel command as they will be sharing the same mqtt connection.

Signed-off-by: yengliong <yeng.liong.wong@intel.com>

* Cancel request in the middle of download

* Fix unit test

* Add unit tests to increase coverage

* Add timeout in thread.join

---------

Signed-off-by: yengliong <yeng.liong.wong@intel.com>