[NEXMANAGE-899] Fix dispatcher not recording state file and causing snapshot failure
The dispatcher agent raises an exception when it fails to retrieve the
state value from the state file. This results in the dispatcher not
recording the restart_reason and snapshot_num, which are important for
the rollback scenario.
This PR removes the exception that occurs when the dispatcher agent is
unable to retrieve the state value from the state file. In cases where
the state is None, the dispatcher will still record the restart_reason and
snapshot_num.

Signed-off-by: yengliong <yeng.liong.wong@intel.com>