[NEXMANAGE-949] Check the IMAGE_BUILD_DATE instead of VERSION for TiberOS SOTA verification
After SOTA system reboot, the VERSION in /etc/os-release remains the same as before.

However, if we check the image build date at /etc/image-id, it's showing the correct image that we want to flash.

Based on the testing, the dispatcher agent should checks the IMAGE_BUILD_DATE instead of VERSION.

Signed-off-by: yengliong <yeng.liong.wong@intel.com>