[NEXMANAGE-826] Update granular log to support arch-independent package
Some packages are architecture-independent and are denoted with the suffix "all", for example, linux-firmware:all. In the history.log, it is recorded as linux-firmware:amd64. However, when using the dpkg command, it appears as linux-firmware:all instead of linux-firmware:amd64. This PR updates the granular log to recheck the package status for "all" architecture if the initial attempt fails to retrieve the package status.

Signed-off-by: yengliong <yeng.liong.wong@intel.com>