[NEX-15262] Returns failure reason that matches the format required by MM
This PR refines the failure_reason in the granular log to match the
mapping required by MM. It detects keywords in the error message and
maps the error message to the corresponding Enum of the failure_reason.

This PR also fixes an issue where the granular log handler didn't record
the command's error when run_commands failed to run the cmd_list.

Signed-off-by: yengliong <yeng.liong.wong@intel.com>